pyproject.toml
[project]
name = "slm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets==3.4.1",
    "huggingface-hub==0.34.0",
    "pyyaml>=6.0.3",
    "reasoning-gym>=0.1.24",
    "rich>=14.3.2",
    "torch>=2.10.0",
    "tqdm>=4.67.3",
    "transformers==4.57.6",
    "trl==0.24.0",
    "unsloth>=2026.1.4",
]
